Fixed the bug that GUIHelper.DrawSerializedProperty didn't update properly so couldn't get the data & Make user can modify spreadsheet and worksheet's name in SheetData.
In Unity 2019.3.15f1 (Not sure about the latest version), GUIHelper.DrawSerializedProperty seems to have some kind of bugs cause the UI can't update.
The targetData in the custon Load() function have gotten the data array, but the UI didn't update and save.
So I deprecated the GUIHelper.DrawSerializedProperty and use Unity's original function "EditorGUILayout.PropertyField".
Aslo the spreadsheet and worksheet's name is the same just use PropertyField to save or modify value.